Fix serious benchmark flaw resulting in unfair comparison.
The benchmark script does not use two of the libraries
in the comparison as they were designed to be used. The
result is that the benchmark reports abysmal performance
for them. As a result, the relative performance as
advertised in the README is incorrect and VERY misleading.

Specifically lines-and-columns and line-column build indexes
on the input string, greatly speeding up position lookups.
The benchmark was creating a new instance of these indexes
EVERY lookup. It should be noted that string-pos uses a cache
of previous calls that acts effectively like an index; thus
string-pos was benefiting from an optimization while the
aforementioned libraries where denied theirs.

The `results.png` displayed on the README is also
updated in this commit.